git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@8775 bc24b658-cf39-4249-b827-e3526eaaeb42

+++ grouper-misc/grouperClient/src/java/edu/internet2/middleware/grouperClient/util/ExpirableCache.java
@@ -101,7 +101,7 @@ public class ExpirableCache<K,V> implements Serializable {
     super();
     GrouperClientUtils.assertion(defaultTimeToLiveInMinutes > 0, "Time to live in minutes must be greater than 0");
     //make sure this is less than the max
-    long newTimeToLiveMillis = defaultTimeToLiveInMinutes * 60 * 1000;
+    long newTimeToLiveMillis = (long)defaultTimeToLiveInMinutes * 60 * 1000;
     if (newTimeToLiveMillis < MAX_TIME_TO_LIVE_MILLIS) {
       this.defaultTimeToLiveInMillis = newTimeToLiveMillis;
     }
@@ -121,7 +121,7 @@ public class ExpirableCache<K,V> implements Serializable {
        */
       @Override
       public long defaultTimeToLiveMillis(int input) {
-        return input * 60 * 1000;
+        return (long)input * 60 * 1000;
       }
     },
     
@@ -133,7 +133,7 @@ public class ExpirableCache<K,V> implements Serializable {
        */
       @Override
       public long defaultTimeToLiveMillis(int input) {
-        return input * 1000;
+        return (long)input * 1000;
       }
     };
     
@@ -193,7 +193,7 @@ public class ExpirableCache<K,V> implements Serializable {
     }
     
     GrouperClientUtils.assertion(timeToLiveInMinutes > 0, "Time to live in minutes must be greater than 0");
-    this.putHelper(key, value, timeToLiveInMinutes * 60 * 1000);
+    this.putHelper(key, value, (long)timeToLiveInMinutes * 60 * 1000);
   }
 
   /**
